BREAKING CHANGE: Use `cy.task` for health check rather than `cy.exec`
This will fix #104 (and hopefully #138).

`cy.task` will always execute within the version of node that is bundled with
Cypress, so we no longer have to worry about `$PATH` issues that `cy.exec`
faces. We're also no longer running a CLI command for the health check, this new
implementation will make a HTTP request from node.

This does mean we need to update the docs to let users know there's an extra
step to configure the SDK properly now.

In an ideal world we would be able to `.catch` on `cy.request` and disable Percy
after the first failed `POST` of the DOM. I think in the future we should look
into working with Cypress to implement this so we can shrink the integration,
make the SDK more reliable, and faster (since we won't make 2 network requests
per-snapshot).