:bug: Increase timeout to 1.5x the asset discovery timeout (30s) (#380)
Assets that take a long time to respond in asset discovery can cause the
`cy.then` promise to timeout. We _must_ wait for all assets to be
discovered/resolved/captured before moving onto the next test, or we risk
capturing the wrong asset for that given snapshot. For example, if the server
shuts down after each test, we would fail to capture those assets if we didn't
await on asset discovery to finish.

With that said, we can safely increase this timeout beyond asset discoveries
timeout since asset discovery will timeout first.